+++ src/nodes/widgets/widgetNodes.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
-
+import * as PIXI from 'pixi.js';
 import React, { useEffect, useRef, useState } from 'react';
 import {
   Button,
@@ -25,8 +25,9 @@ import {
 import ColorizeIcon from '@mui/icons-material/Colorize';
 import { SketchPicker } from 'react-color';
 import Socket from '../../classes/SocketClass';
-import { Widget_Base } from './abstract';
+import { Widget_Base, WidgetHybridBase } from './abstract';
 import { TRgba } from '../../utils/interfaces';
+import { limitRange } from '../../utils/utils';
 import {
   PRESET_COLORS,
   RANDOMMAINCOLOR,
@@ -45,7 +46,6 @@ const selectedName = 'Initial selection';
 const initialValueName = 'Initial value';
 const minValueName = 'Min';
 const roundName = 'Round';
-const stepSizeName = 'Step size';
 const maxValueName = 'Max';
 const offValueName = 'Off';
 const onValueName = 'On';
@@ -59,7 +59,33 @@ const margin = 4;
 
 const defaultOptions = ['Option1', 'Option2', 'Option3'];
 
-export class WidgetButton2 extends Widget_Base {
+const fillWhiteHex = TRgba.white().hex();
+const fillColorDarkHex = TRgba.fromString(RANDOMMAINCOLOR).darken(0.5).hex();
+const fillColorHex = TRgba.fromString(RANDOMMAINCOLOR).hex();
+const contrastColorHex = TRgba.fromString(RANDOMMAINCOLOR)
+  .getContrastTextColor()
+  .hex();
+
+const baseStyle = {
+  fontFamily: ['Roboto', 'Helvetica', 'Arial', 'sans-serif'],
+  fontSize: 16,
+  letterSpacing: 0.45,
+  fill: contrastColorHex,
+  wordWrap: true,
+};
+
+export class WidgetButton extends Widget_Base {
+  _refLabel: PIXI.Text;
+  _refWidget: Button;
+  _refGraphics: PIXI.Graphics;
+
+  private labelTextStyle = new PIXI.TextStyle({
+    ...baseStyle,
+    align: 'center',
+    fontWeight: '500',
+    fill: contrastColorHex,
+  });
+
   protected getUpdateBehaviour(): UpdateBehaviourClass {
     return new UpdateBehaviourClass(false, false, 1000);
   }
@@ -74,7 +100,7 @@ export class WidgetButton2 extends Widget_Base {
   }
 
   public getName(): string {
-    return 'Button';
+    return 'Button2';
   }
 
   public getDescription(): string {
@@ -89,81 +115,85 @@ export class WidgetButton2 extends Widget_Base {
     return 104;
   }
 
+  handleOnPointerDown = () => {
+    this.onWidgetTrigger();
+
+    this._refWidget.view.scale.x = 0.99;
+    this._refWidget.view.scale.y = 0.99;
+    this._refWidget.view.alpha = 0.8;
+    const inputData = this.getInputData(onValueName);
+    this.setOutputData(outName, inputData);
+    this.executeChildren();
+  };
+
+  handleOnPointerUp = () => {
+    this._refWidget.view.scale.x = 1;
+    this._refWidget.view.scale.y = 1;
+    this._refWidget.view.alpha = 1;
+    const inputData = this.getInputData(offValueName);
+    this.setOutputData(outName, inputData);
+    this.executeChildren();
+  };
+
   public onWidgetTrigger = () => {
     console.log('onWidgetTrigger');
     this.executeOptimizedChain();
   };
 
-  protected getParentComponent(props: any): any {
-    const node = props.node;
+  public onNodeAdded = (source?: TNodeSource) => {
+    this._refGraphics = new PIXI.Graphics()
+      .beginFill(fillColorHex)
+      .drawRoundedRect(
+        0,
+        0,
+        this.nodeWidth - 8 * margin,
+        this.nodeHeight - 8 * margin,
+        16
+      );
+    this._refWidget = new Button(this._refGraphics);
 
-    useEffect(() => {
-      node.setOutputData(outName, node.getInputData(offValueName));
-      node.executeChildren();
-    }, []);
+    this._refGraphics.pivot.x = 0;
+    this._refGraphics.pivot.y = 0;
+    this._refWidget.view.x = NODE_MARGIN + 4 * margin;
+    this._refWidget.view.y = 4 * margin;
+    this._refWidget.onDown.connect(this.handleOnPointerDown);
+    this._refWidget.onUp.connect(this.handleOnPointerUp);
+    this.addChild(this._refWidget.view);
 
-    const handleOnPointerDown = () => {
-      node.onWidgetTrigger();
-      const inputData = node.getInputData(onValueName);
-      node.setOutputData(outName, inputData);
-      node.executeChildren();
-    };
-
-    const handleOnPointerUp = () => {
-      const inputData = node.getInputData(offValueName);
-      node.setOutputData(outName, inputData);
-      node.executeChildren();
-    };
-
-    return (
-      <ThemeProvider theme={customTheme}>
-        <Paper
-          component={Stack}
-          direction="column"
-          justifyContent="center"
-          sx={{
-            bgcolor: 'background.default',
-            fontSize: '16px',
-            border: 0,
-            width: `${node.nodeWidth}px`,
-            height: `${node.nodeHeight}px`,
-            boxShadow: 16,
-            '&:hover': {
-              boxShadow: 12,
-            },
-          }}
-        >
-          <Button
-            variant="contained"
-            onPointerDown={handleOnPointerDown}
-            onPointerUp={handleOnPointerUp}
-            sx={{
-              pointerEvents: 'auto',
-              margin: 'auto',
-              fontSize: '16px',
-              lineHeight: '20px',
-              border: 0,
-              width: `${node.nodeWidth - 8 * margin}px`,
-              height: `${node.nodeHeight - 8 * margin}px`,
-              borderRadius: `${node.nodeWidth / 16}px`,
-              boxShadow: 16,
-              '&:hover': {
-                boxShadow: 12,
-              },
-              '&:active': {
-                boxShadow: 4,
-              },
-            }}
-          >
-            {props[labelName]}
-          </Button>
-        </Paper>
-      </ThemeProvider>
+    this._refLabel = new PIXI.Text(
+      String(this.getInputData(labelName)).toUpperCase(),
+      this.labelTextStyle
     );
-  }
+    this._refLabel.anchor.x = 0.5;
+    this._refLabel.anchor.y = 0.5;
+    this._refLabel.style.wordWrapWidth = this.nodeWidth - 10 * margin;
+    this._refLabel.x = NODE_MARGIN + this.nodeWidth / 2;
+    this._refLabel.y = this.nodeHeight / 2;
+    this._refLabel.eventMode = 'none';
+    this.addChild(this._refLabel);
+
+    super.onNodeAdded(source);
+  };
+
+  public onNodeResize = (newWidth, newHeight) => {
+    this._refGraphics.clear();
+    this._refGraphics
+      .beginFill(fillColorHex)
+      .drawRoundedRect(0, 0, newWidth - 8 * margin, newHeight - 8 * margin, 16);
+    this._refWidget.view.width = newWidth - 8 * margin;
+    this._refWidget.view.height = newHeight - 8 * margin;
+    this._refLabel.x = NODE_MARGIN + newWidth / 2;
+    this._refLabel.y = newHeight / 2;
+    this._refLabel.style.wordWrapWidth = newWidth - 10 * margin;
+  };
+
+  public onExecute = async (input, output) => {
+    const text = String(input[labelName]).toUpperCase();
+    this._refLabel.text = text;
+  };
 }
 
-export class WidgetColorPicker extends Widget_Base {
+export class WidgetColorPicker extends WidgetHybridBase {
   protected getDefaultIO(): Socket[] {
     return [
       new Socket(
@@ -309,7 +339,7 @@ export class WidgetColorPicker extends Widget_Base {
   }
 }
 
-export class WidgetSwitch extends Widget_Base {
+export class WidgetSwitch extends WidgetHybridBase {
   protected getDefaultIO(): Socket[] {
     return [
       new Socket(SOCKET_TYPE.IN, selectedName, new BooleanType(), false, false),
@@ -409,25 +439,49 @@ export class WidgetSwitch extends Widget_Base {
   }
 }
 
-export class WidgetSlider2 extends Widget_Base {
+export class WidgetSlider extends Widget_Base {
+  _refLabel: PIXI.Text;
+  _refValue: PIXI.Text;
+  _refWidget: Slider;
+  _refBg: PIXI.Graphics;
+  _refFill: PIXI.Graphics;
+  _refSlider: PIXI.Graphics;
+
+  private valueTextStyle = new PIXI.TextStyle({
+    ...baseStyle,
+    align: 'center',
+    fontWeight: '500',
+    fill: fillWhiteHex,
+  });
+
+  private labelTextStyle = new PIXI.TextStyle({
+    ...baseStyle,
+    align: 'center',
+    fontWeight: '100',
+    fill: fillWhiteHex,
+  });
+
+  // protected getUpdateBehaviour(): UpdateBehaviourClass {
+  //   return new UpdateBehaviourClass(false, false, 1000);
+  // }
+
   protected getDefaultIO(): Socket[] {
     return [
       new Socket(SOCKET_TYPE.IN, initialValueName, new NumberType(), 0, false),
       new Socket(SOCKET_TYPE.IN, minValueName, new NumberType(), 0, false),
       new Socket(SOCKET_TYPE.IN, maxValueName, new NumberType(), 100, false),
       new Socket(SOCKET_TYPE.IN, roundName, new BooleanType(), 100, false),
-      new Socket(SOCKET_TYPE.IN, stepSizeName, new NumberType(), 0.01, false),
       new Socket(SOCKET_TYPE.IN, labelName, new StringType(), 'Slider', false),
       new Socket(SOCKET_TYPE.OUT, outName, new NumberType()),
     ];
   }
 
   public getName(): string {
-    return 'Slider-old';
+    return 'Slider';
   }
 
   public getDescription(): string {
-    return 'Adds a number slider';
+    return 'Adds a button to trigger values';
   }
 
   public getDefaultNodeWidth(): number {
@@ -438,137 +492,149 @@ export class WidgetSlider2 extends Widget_Base {
     return 104;
   }
 
-  protected getParentComponent(props: any): any {
-    const node = props.node;
-    const [data, setData] = useState(Number(props[initialValueName]));
-    const [minValue, setMinValue] = useState(
-      Math.min(props[minValueName] ?? 0, data)
+  public onNodeAdded = (source?: TNodeSource) => {
+    // Widget
+    this._refBg = new PIXI.Graphics()
+      .beginFill(fillColorDarkHex)
+      .drawRoundedRect(
+        0,
+        0,
+        this.nodeWidth - 8 * margin,
+        this.nodeHeight - 16 * margin,
+        16
+      );
+
+    this._refFill = new PIXI.Graphics()
+      .beginFill(fillColorHex)
+      .drawRoundedRect(
+        0,
+        0,
+        this.nodeWidth - 8 * margin,
+        this.nodeHeight - 16 * margin,
+        16
+      );
+
+    this._refSlider = new PIXI.Graphics();
+
+    this._refWidget = new Slider({
+      bg: this._refBg,
+      fill: this._refFill,
+      slider: this._refSlider,
+      min: this.getInputData(minValueName),
+      max: this.getInputData(maxValueName),
+      value: this.getInputData(initialValueName),
+      valueTextStyle: this.valueTextStyle,
+      showValue: false,
+    });
+
+    this._refWidget.x = NODE_MARGIN + 4 * margin;
+    this._refWidget.y = 4 * margin;
+    this._refWidget.onUpdate.connect(this.handleOnChange);
+
+    this.addChild(this._refWidget);
+
+    // Label
+    this._refLabel = new PIXI.Text(
+      String(this.getInputData(labelName)),
+      this.labelTextStyle
     );
-    const [maxValue, setMaxValue] = useState(
-      Math.max(props[maxValueName] ?? 100, data)
+    this._refLabel.anchor.x = 0.5;
+    this._refLabel.anchor.y = 1;
+    this._refLabel.style.wordWrapWidth = this.nodeWidth - 10 * margin;
+    this._refLabel.x = NODE_MARGIN + this.nodeWidth / 2;
+    this._refLabel.y = this.nodeHeight - 2 * margin;
+    this._refLabel.eventMode = 'none';
+    this.addChild(this._refLabel);
+
+    // Value
+    this._refValue = new PIXI.Text(
+      String(this.getInputData(initialValueName)),
+      this.valueTextStyle
     );
-    const [round, setRound] = useState(props[roundName] ?? false);
-    const [stepSizeValue, setStepSizeValue] = useState(
-      props[stepSizeName] ?? 0.01
+    this._refValue.anchor.x = 0.5;
+    this._refValue.anchor.y = 0;
+    this._refValue.x = NODE_MARGIN + this.nodeWidth / 2;
+    this._refValue.y = 2 * margin;
+    this._refValue.eventMode = 'none';
+    this.addChild(this._refValue);
+
+    super.onNodeAdded(source);
+  };
+
+  public onNodeResize = (newWidth, newHeight) => {
+    this._refWidget.progress = this.valueToPercent(
+      this.getInputData(initialValueName)
     );
 
-    useEffect(() => {
-      node.setOutputData(outName, data);
-      node.executeChildren();
-    }, []);
+    this._refBg.clear();
+    this._refBg
+      .beginFill(fillColorDarkHex)
+      .drawRoundedRect(
+        0,
+        0,
+        newWidth - 8 * margin,
+        newHeight - 16 * margin,
+        16
+      );
 
-    useEffect(() => {
-      setData(Number(props[initialValueName]));
-      setMinValue(Math.min(props[minValueName] ?? 0, data));
-      setMaxValue(Math.max(props[maxValueName] ?? 100, data));
-      setRound(props[roundName] ?? false);
-      setStepSizeValue(props[stepSizeName] ?? 0.01);
-    }, [
-      props[initialValueName],
-      props[minValueName],
-      props[maxValueName],
-      props[roundName],
-      props[stepSizeName],
-    ]);
+    this._refFill.clear();
+    this._refFill
+      .beginFill(fillColorHex)
+      .drawRoundedRect(
+        0,
+        0,
+        newWidth - 8 * margin,
+        newHeight - 16 * margin,
+        16
+      );
 
-    const handleOnChange = (event, value) => {
-      if (!Array.isArray(value)) {
-        if (value !== props[initialValueName]) {
-          node.setInputData(initialValueName, value);
-        }
-        setData(value);
-        node.setOutputData(outName, value);
-        node.executeChildren();
-      }
-    };
+    this._refWidget.y = (newHeight - (newHeight - 16 * margin)) / 2;
 
-    return (
-      <ThemeProvider theme={customTheme}>
-        <Paper
-          component={Stack}
-          direction="column"
-          justifyContent="center"
-          sx={{
-            bgcolor: 'background.default',
-            fontSize: '16px',
-            border: 0,
-            width: `${node.nodeWidth}px`,
-            height: `${node.nodeHeight}px`,
-            boxShadow: 16,
-            '&:hover': {
-              boxShadow: 12,
-            },
-          }}
-        >
-          <Slider
-            size="small"
-            color="secondary"
-            valueLabelDisplay="on"
-            min={minValue}
-            max={maxValue}
-            step={round ? 1 : stepSizeValue}
-            onChange={handleOnChange}
-            value={data || 0}
-            sx={{
-              margin: `${6 * margin}px ${8 * margin}px ${margin}px`,
-              height: '8px',
-              pointerEvents: 'auto',
-              '&.MuiSlider-root': {
-                width: 'unset',
-              },
-              '& .MuiSlider-rail': {
-                backgroundColor: 'primary.dark',
-              },
-              '& .MuiSlider-track': {
-                border: 'none',
-                backgroundColor: 'primary.main',
-              },
-              '& .MuiSlider-valueLabel': {
-                fontSize: '1rem',
-                fontWeight: 'normal',
-                top: -4,
-                backgroundColor: 'unset',
-                color: 'text.primary',
-                '&:before': {
-                  display: 'none',
-                },
-                '& *': {
-                  background: 'transparent',
-                  color: 'text.primary',
-                },
-              },
-              '& .MuiSlider-thumb': {
-                height: 32,
-                width: 32,
-                backgroundColor: 'text.primary',
-                borderColor: 'primary.main',
-                borderWidth: '4px',
-                borderStyle: 'solid',
-                '&:focus, &:hover, &.Mui-active, &.Mui-focusVisible': {
-                  boxShadow: 'inherit',
-                },
-                '&:before': {
-                  display: 'none',
-                },
-              },
-            }}
-          />
-          <Typography
-            sx={{
-              textAlign: 'center',
-              textOverflow: 'ellipsis',
-              px: 2,
-            }}
-          >
-            {props[labelName]}
-          </Typography>
-        </Paper>
-      </ThemeProvider>
-    );
-  }
+    this._refValue.x = NODE_MARGIN + newWidth / 2;
+
+    this._refLabel.x = NODE_MARGIN + newWidth / 2;
+    this._refLabel.y = this.nodeHeight - 2 * margin;
+    this._refLabel.style.wordWrapWidth = newWidth - 10 * margin;
+  };
+
+  valueToPercent = (value) => {
+    const minValue = this.getInputData(minValueName);
+    const maxValue = this.getInputData(maxValueName);
+    return ((value - minValue) / (maxValue - minValue)) * 100;
+  };
+
+  setOutputDataAndText = (value) => {
+    const shouldRound = this.getInputData(roundName);
+    const newValue = shouldRound ? Math.round(value) : value;
+    this._refValue.text = roundNumber(newValue, shouldRound ? 0 : 2);
+    this.setOutputData(outName, newValue);
+  };
+
+  handleOnChange = (value) => {
+    this.setInputData(initialValueName, value);
+    this.setOutputDataAndText(value);
+    this.executeChildren();
+  };
+
+  public onExecute = async (input, output) => {
+    const value = input[initialValueName];
+    const minValue = input[minValueName];
+    const maxValue = input[maxValueName];
+    this._refWidget.min = minValue;
+    this._refWidget.max = maxValue;
+
+    const text = String(input[labelName]);
+    this._refLabel.text = text;
+
+    // update the slider in percent
+    this._refWidget.progress = this.valueToPercent(value);
+
+    // update the output
+    this.setOutputDataAndText(limitRange(value, minValue, maxValue));
+  };
 }
 
-export class WidgetDropdown extends Widget_Base {
+export class WidgetDropdown extends WidgetHybridBase {
   protected getDefaultIO(): Socket[] {
     return [
       new Socket(
