--- src/nodes/widgets/widgetNodes2.tsx
@@ -1,369 +1,0 @@
-/* eslint-disable @typescript-eslint/no-empty-function */
-import * as PIXI from 'pixi.js';
-import { Button, Slider } from '@pixi/ui';
-
-import Socket from '../../classes/SocketClass';
-import { Widget_Base2 } from './abstract';
-import {
-  NODE_MARGIN,
-  RANDOMMAINCOLOR,
-  SOCKET_TYPE,
-} from '../../utils/constants';
-import { limitRange, roundNumber } from '../../utils/utils';
-import { TNodeSource, TRgba } from '../../utils/interfaces';
-import { AnyType } from '../datatypes/anyType';
-import { ArrayType } from '../datatypes/arrayType';
-import { BooleanType } from '../datatypes/booleanType';
-import { NumberType } from '../datatypes/numberType';
-import { StringType } from '../datatypes/stringType';
-import { ColorType } from '../datatypes/colorType';
-import UpdateBehaviourClass from '../../classes/UpdateBehaviourClass';
-
-const selectedName = 'Initial selection';
-const initialValueName = 'Initial value';
-const minValueName = 'Min';
-const roundName = 'Round';
-const maxValueName = 'Max';
-const offValueName = 'Off';
-const onValueName = 'On';
-const labelName = 'Label';
-const optionsName = 'Options';
-const selectedOptionName = 'Selected option';
-const multiSelectName = 'Select multiple';
-const outName = 'Out';
-
-const margin = 4;
-
-const defaultOptions = ['Option1', 'Option2', 'Option3'];
-
-// const fillColorDarkHex = TRgba.fromString(Color(RANDOMMAINCOLOR).darken(0.85).hex());
-const fillWhiteHex = TRgba.white().hex();
-const fillColorDarkHex = TRgba.fromString(RANDOMMAINCOLOR).darken(0.5).hex();
-const fillColorHex = TRgba.fromString(RANDOMMAINCOLOR).hex();
-const contrastColorHex = TRgba.fromString(RANDOMMAINCOLOR)
-  .getContrastTextColor()
-  .hex();
-
-const baseStyle = {
-  fontFamily: ['Roboto', 'Helvetica', 'Arial', 'sans-serif'],
-  fontSize: 16,
-  letterSpacing: 0.45,
-  fill: contrastColorHex,
-  wordWrap: true,
-};
-
-export class WidgetButton extends Widget_Base2 {
-  _refLabel: PIXI.Text;
-  _refWidget: Button;
-  _refGraphics: PIXI.Graphics;
-
-  private labelTextStyle = new PIXI.TextStyle({
-    ...baseStyle,
-    align: 'center',
-    fontWeight: '500',
-    fill: contrastColorHex,
-  });
-
-  protected getUpdateBehaviour(): UpdateBehaviourClass {
-    return new UpdateBehaviourClass(false, false, 1000);
-  }
-
-  protected getDefaultIO(): Socket[] {
-    return [
-      new Socket(SOCKET_TYPE.IN, offValueName, new AnyType(), 0, false),
-      new Socket(SOCKET_TYPE.IN, onValueName, new AnyType(), 1, false),
-      new Socket(SOCKET_TYPE.IN, labelName, new StringType(), 'Button', false),
-      new Socket(SOCKET_TYPE.OUT, outName, new AnyType()),
-    ];
-  }
-
-  public getName(): string {
-    return 'Button2';
-  }
-
-  public getDescription(): string {
-    return 'Adds a button to trigger values';
-  }
-
-  public getDefaultNodeWidth(): number {
-    return 200;
-  }
-
-  public getDefaultNodeHeight(): number {
-    return 104;
-  }
-
-  handleOnPointerDown = () => {
-    this.onWidgetTrigger();
-
-    this._refWidget.view.scale.x = 0.99;
-    this._refWidget.view.scale.y = 0.99;
-    this._refWidget.view.alpha = 0.8;
-    const inputData = this.getInputData(onValueName);
-    this.setOutputData(outName, inputData);
-    this.executeChildren();
-  };
-
-  handleOnPointerUp = () => {
-    this._refWidget.view.scale.x = 1;
-    this._refWidget.view.scale.y = 1;
-    this._refWidget.view.alpha = 1;
-    const inputData = this.getInputData(offValueName);
-    this.setOutputData(outName, inputData);
-    this.executeChildren();
-  };
-
-  public onWidgetTrigger = () => {
-    console.log('onWidgetTrigger');
-    this.executeOptimizedChain();
-  };
-
-  public onNodeAdded = (source?: TNodeSource) => {
-    this._refGraphics = new PIXI.Graphics()
-      .beginFill(fillColorHex)
-      .drawRoundedRect(
-        0,
-        0,
-        this.nodeWidth - 8 * margin,
-        this.nodeHeight - 8 * margin,
-        16
-      );
-    this._refWidget = new Button(this._refGraphics);
-
-    this._refGraphics.pivot.x = 0;
-    this._refGraphics.pivot.y = 0;
-    this._refWidget.view.x = NODE_MARGIN + 4 * margin;
-    this._refWidget.view.y = 4 * margin;
-    this._refWidget.onDown.connect(this.handleOnPointerDown);
-    this._refWidget.onUp.connect(this.handleOnPointerUp);
-    this.addChild(this._refWidget.view);
-
-    this._refLabel = new PIXI.Text(
-      String(this.getInputData(labelName)).toUpperCase(),
-      this.labelTextStyle
-    );
-    this._refLabel.anchor.x = 0.5;
-    this._refLabel.anchor.y = 0.5;
-    this._refLabel.style.wordWrapWidth = this.nodeWidth - 10 * margin;
-    this._refLabel.x = NODE_MARGIN + this.nodeWidth / 2;
-    this._refLabel.y = this.nodeHeight / 2;
-    this._refLabel.eventMode = 'none';
-    this.addChild(this._refLabel);
-
-    super.onNodeAdded(source);
-  };
-
-  public onNodeResize = (newWidth, newHeight) => {
-    this._refGraphics.clear();
-    this._refGraphics
-      .beginFill(fillColorHex)
-      .drawRoundedRect(0, 0, newWidth - 8 * margin, newHeight - 8 * margin, 16);
-    this._refWidget.view.width = newWidth - 8 * margin;
-    this._refWidget.view.height = newHeight - 8 * margin;
-    this._refLabel.x = NODE_MARGIN + newWidth / 2;
-    this._refLabel.y = newHeight / 2;
-    this._refLabel.style.wordWrapWidth = newWidth - 10 * margin;
-  };
-
-  public onExecute = async (input, output) => {
-    const text = String(input[labelName]).toUpperCase();
-    this._refLabel.text = text;
-  };
-}
-
-export class WidgetSlider extends Widget_Base2 {
-  _refLabel: PIXI.Text;
-  _refValue: PIXI.Text;
-  _refWidget: Slider;
-  _refBg: PIXI.Graphics;
-  _refFill: PIXI.Graphics;
-  _refSlider: PIXI.Graphics;
-
-  private valueTextStyle = new PIXI.TextStyle({
-    ...baseStyle,
-    align: 'center',
-    fontWeight: '500',
-    fill: fillWhiteHex,
-  });
-
-  private labelTextStyle = new PIXI.TextStyle({
-    ...baseStyle,
-    align: 'center',
-    fontWeight: '100',
-    fill: fillWhiteHex,
-  });
-
-  // protected getUpdateBehaviour(): UpdateBehaviourClass {
-  //   return new UpdateBehaviourClass(false, false, 1000);
-  // }
-
-  protected getDefaultIO(): Socket[] {
-    return [
-      new Socket(SOCKET_TYPE.IN, initialValueName, new NumberType(), 0, false),
-      new Socket(SOCKET_TYPE.IN, minValueName, new NumberType(), 0, false),
-      new Socket(SOCKET_TYPE.IN, maxValueName, new NumberType(), 100, false),
-      new Socket(SOCKET_TYPE.IN, roundName, new BooleanType(), 100, false),
-      new Socket(SOCKET_TYPE.IN, labelName, new StringType(), 'Slider', false),
-      new Socket(SOCKET_TYPE.OUT, outName, new NumberType()),
-    ];
-  }
-
-  public getName(): string {
-    return 'Slider';
-  }
-
-  public getDescription(): string {
-    return 'Adds a button to trigger values';
-  }
-
-  public getDefaultNodeWidth(): number {
-    return 200;
-  }
-
-  public getDefaultNodeHeight(): number {
-    return 104;
-  }
-
-  public onNodeAdded = (source?: TNodeSource) => {
-    // Widget
-    this._refBg = new PIXI.Graphics()
-      .beginFill(fillColorDarkHex)
-      .drawRoundedRect(
-        0,
-        0,
-        this.nodeWidth - 8 * margin,
-        this.nodeHeight - 16 * margin,
-        16
-      );
-
-    this._refFill = new PIXI.Graphics()
-      .beginFill(fillColorHex)
-      .drawRoundedRect(
-        0,
-        0,
-        this.nodeWidth - 8 * margin,
-        this.nodeHeight - 16 * margin,
-        16
-      );
-
-    this._refSlider = new PIXI.Graphics();
-
-    this._refWidget = new Slider({
-      bg: this._refBg,
-      fill: this._refFill,
-      slider: this._refSlider,
-      min: this.getInputData(minValueName),
-      max: this.getInputData(maxValueName),
-      value: this.getInputData(initialValueName),
-      valueTextStyle: this.valueTextStyle,
-      showValue: false,
-    });
-
-    this._refWidget.x = NODE_MARGIN + 4 * margin;
-    this._refWidget.y = 4 * margin;
-    this._refWidget.onUpdate.connect(this.handleOnChange);
-
-    this.addChild(this._refWidget);
-
-    // Label
-    this._refLabel = new PIXI.Text(
-      String(this.getInputData(labelName)).toUpperCase(),
-      this.labelTextStyle
-    );
-    this._refLabel.text = String(this.getInputData(labelName)).toUpperCase();
-    this._refLabel.anchor.x = 0.5;
-    this._refLabel.anchor.y = 1;
-    this._refLabel.style.wordWrapWidth = this.nodeWidth - 10 * margin;
-    this._refLabel.x = NODE_MARGIN + this.nodeWidth / 2;
-    this._refLabel.y = this.nodeHeight - 2 * margin;
-    this._refLabel.eventMode = 'none';
-    this.addChild(this._refLabel);
-
-    // Value
-    this._refValue = new PIXI.Text(
-      String(this.getInputData(labelName)).toUpperCase(),
-      this.valueTextStyle
-    );
-    this._refValue.text = String(this.getInputData(initialValueName));
-    this._refValue.anchor.x = 0.5;
-    this._refValue.anchor.y = 0;
-    this._refValue.x = NODE_MARGIN + this.nodeWidth / 2;
-    this._refValue.y = 2 * margin;
-    this._refValue.eventMode = 'none';
-    this.addChild(this._refValue);
-
-    super.onNodeAdded(source);
-  };
-
-  public onNodeResize = (newWidth, newHeight) => {
-    this._refWidget.progress = this.valueToPercent(
-      this.getInputData(initialValueName)
-    );
-
-    this._refBg.clear();
-    this._refBg
-      .beginFill(fillColorDarkHex)
-      .drawRoundedRect(
-        0,
-        0,
-        newWidth - 8 * margin,
-        newHeight - 16 * margin,
-        16
-      );
-
-    this._refFill.clear();
-    this._refFill
-      .beginFill(fillColorHex)
-      .drawRoundedRect(
-        0,
-        0,
-        newWidth - 8 * margin,
-        newHeight - 16 * margin,
-        16
-      );
-
-    this._refWidget.y = (newHeight - (newHeight - 16 * margin)) / 2;
-
-    this._refValue.x = NODE_MARGIN + newWidth / 2;
-
-    this._refLabel.x = NODE_MARGIN + newWidth / 2;
-    this._refLabel.y = this.nodeHeight - 2 * margin;
-    this._refLabel.style.wordWrapWidth = newWidth - 10 * margin;
-  };
-
-  valueToPercent = (value) => {
-    const minValue = this.getInputData(minValueName);
-    const maxValue = this.getInputData(maxValueName);
-    return ((value - minValue) / (maxValue - minValue)) * 100;
-  };
-
-  setOutputDataAndText = (value) => {
-    const shouldRound = this.getInputData(roundName);
-    const newValue = shouldRound ? Math.round(value) : value;
-    this._refValue.text = roundNumber(newValue, shouldRound ? 0 : 2);
-    this.setOutputData(outName, newValue);
-  };
-
-  handleOnChange = (value) => {
-    this.setInputData(initialValueName, value);
-    this.setOutputDataAndText(value);
-    this.executeChildren();
-  };
-
-  public onExecute = async (input, output) => {
-    const value = input[initialValueName];
-    const minValue = input[minValueName];
-    const maxValue = input[maxValueName];
-    this._refWidget.min = minValue;
-    this._refWidget.max = maxValue;
-
-    const text = String(input[labelName]).toUpperCase();
-    this._refLabel.text = text;
-
-    // update the slider in percent
-    this._refWidget.progress = this.valueToPercent(value);
-
-    // update the output
-    this.setOutputDataAndText(limitRange(value, minValue, maxValue));
-  };
-}
